feat: create a new feature (#28)
* feat: create a new feature

* fix: add missing file
