+++ ini.flake8
@@ -1,0 +1,4 @@
+[flake8]
+max-line-length = 79
+ignore = E203, W503
+exclude = .git,__pycache__,venv
